% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crude.R
\name{crude}
\alias{crude}
\title{crude}
\usage{
crude(deaths, population, perPopulation, decimal)
}
\arguments{
\item{deaths}{Vector of death counts for each age category}

\item{population}{Vector of population counts for each age category}

\item{perPopulation}{Optional input for "per population" that applies to rate (default: per 1000 members of population)}

\item{decimal}{Optional input for rounding rate to certain decimal (default: two decimals)}
}
\value{
Character string including crude mortality rate and per population
}
\description{
Calculates crude mortality rate
}
