% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directAA.R
\name{directAA}
\alias{directAA}
\title{directAA}
\usage{
directAA(deaths, population, refPopulation, perPopulation, decimal)
}
\arguments{
\item{deaths}{Vector of death counts for each age category}

\item{population}{Vector of population counts for each age category}

\item{refPopulation}{Vector of standard population counts for each age category}

\item{perPopulation}{Optional input for "per population" that applies to rate (default: per 1000 members of population)}

\item{decimal}{Optional input for rounding rate to certain decimal (default: two decimals)}
}
\value{
DAR: character string including age-adjusted mortality rate, method of adjustment (direct), and per population

AS: numeric vector of age-specific mortality rates
}
\description{
Calculates age-adjusted mortality rate using the direct method
}
