% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indirectAA.R
\name{indirectAA}
\alias{indirectAA}
\title{indirectAA}
\usage{
indirectAA(
  deaths,
  population,
  refDeathRates,
  refMortality,
  perPopulation,
  decimal
)
}
\arguments{
\item{deaths}{Vector of death counts for each age category}

\item{population}{Vector of population counts for each age category}

\item{refDeathRates}{Vector of standard population death rates for each age category}

\item{refMortality}{Single integer describing overall mortality in reference population}

\item{perPopulation}{Optional input for "per population" that applies to rate (default: per 1000 members of population)}

\item{decimal}{Optional input for rounding rate to certain decimal (default: two decimals)}
}
\value{
IAR: character string including age-adjusted mortality rate, method of adjustment (indirect), and per population

AS: numeric vector of age-specific mortality rates
}
\description{
Calculates age-adjusted mortality rate using the indirect method
}
