% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_table.R
\name{life_table}
\alias{life_table}
\title{life_table}
\usage{
life_table(age, population, deaths)
}
\arguments{
\item{age}{Vector of age categories}

\item{population}{Vector of population counts for each age category}

\item{deaths}{Vector of death counts for each age category}
}
\value{
Data frame of life table consisting of the following columns: age, population, deaths, cpd: conditional probability of death, cps: conditional probability of survival, survival_prop: survival proportion, person_years, total_years: total person years, life_expectancy
}
\description{
Calculates and outputs life table
}
