% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_survival.R
\name{median_survival}
\alias{median_survival}
\title{median_survival}
\usage{
median_survival(lifeTable)
}
\arguments{
\item{lifeTable}{Input must be data frame with columns ages and survival_prop}
}
\value{
Median survival time (in units of age groups that data is input in)
}
\description{
Calculates median survival time
}
