% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_plot.R
\name{survival_plot}
\alias{survival_plot}
\title{survival_plot}
\usage{
survival_plot(lifeTable)
}
\arguments{
\item{lifeTable}{Input must be data frame with columns ages and survival_prop}
}
\value{
step plot of survival probability versus age
}
\description{
Outputs plot of survival probability versus age
}
